Use a fixed version of Node (currently v8.5.0)
It is important we use the same Node version both for
building inside Docker and on the AMI for EC2.
Otherwise, we risk crashes due to incompatible versions
as it currently happens with Node v8.5.0 and v8.6.0.

To install Node, we now download Node directly from nodejs.org.
Installation through PPA does not work, as they do not
support to install older versions.