Making the tests pass on Node 14. For details, see the comments in the code.
To dodge the bug, I changed new Uint8Array() to Uint8Array(1) at
some places. Also, I disabled the part that will trigger it, but
left it enable on non-native builds (wasm and asmjs). For documentation,
I added a simplified version of the affected test to create awareness
of the issue. Production should not be affected, but if you make changes
be careful not to trigger the bug.

Perhaps upgrading the whole tool chain will eventually solve it.

Fixes #14